+++ modules/flake-parts/checks.nix
@@ -1,0 +1,10 @@
+{
+  perSystem = { self', lib, ... }: {
+    checks = let
+      packages =
+        lib.mapAttrs' (n: lib.nameValuePair "package-${n}") self'.packages;
+      devShells =
+        lib.mapAttrs' (n: lib.nameValuePair "devShell-${n}") self'.devShells;
+    in packages // devShells;
+  };
+}