+++ modules/flake-parts/flake-module.nix
@@ -1,0 +1,12 @@
+{
+  imports = [
+    ./checks.nix
+    ./clan.nix
+    ./configurations.nix
+    ./devShells.nix
+    ./lib.nix
+    ./terraform.nix
+    ./treefmt.nix
+    ./vm.nix
+  ];
+}