+++ modules/flake-parts/lib.nix
@@ -1,0 +1,1 @@
+{ self-lib, ... }: { flake.lib = self-lib; }