+++ modules/flake-parts/terraform.nix
@@ -1,0 +1,8 @@
+{
+  perSystem = { pkgs, ... }: {
+    packages.terraform = pkgs.opentofu.withPlugins (p:
+      builtins.attrValues {
+        inherit (p) external hcloud local null onepassword tailscale;
+      });
+  };
+}